fpga: Add the kernel modules (#119)
For the `uio` devices (to access the FPGA memory from Linux) and the ROM
backdoors (for injecting the ROMs into the FPGA).

This is modified a bit from the standard Caliptra backdoor so that we
can define an additional backdoor for the MCU ROM.

These can be build and installed with `cargo xtask
fpga-install-kernel-modules`.

The ROM backdoor devices are not used in the current Caliptra tests, but
will be useful for MCU because we are running out of `uio` segments in
our `uio` device (we're limited to 5).

These modules were tested on the Caliptra Core (2.0) FPGA -- the smoke
test passes, and I was able to dump the Caliptra ROM.

+++ emulator/periph/src/flash_ctrl.rs
@@ -774,7 +774,7 @@ mod test {
         data: Option<&[u8]>,
     ) -> Option<u32> {
         // Check if ref_addr is within the range of DCCM
-        if ref_addr < RAM_OFFSET || ref_addr + size as u32 > RAM_OFFSET + RAM_SIZE as u32 {
+        if ref_addr < RAM_OFFSET || ref_addr + size as u32 > RAM_OFFSET + RAM_SIZE {
             return None;
         }
 
@@ -1421,7 +1421,7 @@ mod test {
         );
     }
 
-    //// TEST CASE STARTED HERE
+    /// TEST CASE STARTED HERE
     #[test]
     fn test_main_flash_regs_access() {
         test_flash_ctrl_regs_access(FlashType::Main);
